added tests

+++ src/VecSim/utils/vec_utils.h
@@ -1,3 +1,4 @@
 #pragma once
+#include "stdlib.h"
 
-void float_vector_normalize(float *x, int dim);
+void float_vector_normalize(float *x, size_t dim);
